Fix totalPages to explicitly cast to int
Ensures ceil() returns an integer rather than a float when totalRows is 0.

+++ www/app/Http/Controllers/MemoryController.php
@@ -254,7 +254,7 @@ class MemoryController extends Controller
                 $processedRows[] = $processedRow;
             }
 
-            $totalPages = ceil($totalRows / $perPage);
+            $totalPages = (int) ceil($totalRows / $perPage);
 
             return view('config.memory-browse', [
                 'tableName' => $tableName,
